+++ changelog.md
@@ -1,7 +1,21 @@
 # ACEmulator Change Log
+
 ### 2017-10-20
 [Mogwai]
 * Authentication overhaul is done.  Please note there are significant changes to the Readme file.
+
+### 2017-10-19
+[Og II]
+* Added processing for contract management add / abandon.
+* Removed update script I added back on 10/14 to fix the burden of the weenie for the MMD note.   This is no lonager needed - use Dereth v0.2.4 or greater the issue has been 
+* resolved in the data and the script is no longer needed.
+* Minor code cleanup and improved some method documentation.
+* Added new table to store persisted tracked contracts (quests)
+* Implemented two new game events / SendClientContractTrackerTable and SendClientContractTracker
+* Implemented abandon quest - Game action - Social_AbandonContract
+* Documeted methods and updated trello.
+* To test @ci 44163 -create a contract.   Use the contract and you are now tracking that quest.   You can create any contract using ci or buy one from a vendor if they are In
+* did not check vendors.   You can abandon the contract.   All management and the cooldown delay of 2 seconds is respected.
 
 ### 2017-10-18
 [StackOverflow]